+++ packages/layout/src/web-components/mlc-layout/types.ts
@@ -45,7 +45,17 @@ export type LocalizedText = string | Record<string, string>
 /** Definition of a dynamically loaded icon */
 export interface Icon {
   /** Library from witch the icon is pulled */
-  library: '@ant-design/icons-svg' | '@fortawesome/free-brands-svg-icons' | '@fortawesome/free-regular-svg-icons' | '@fortawesome/free-solid-svg-icons'
+  library:
+    | '@ant-design/icons-svg'
+    | '@fortawesome/free-brands-svg-icons'
+    | '@fortawesome/free-regular-svg-icons'
+    | '@fortawesome/free-solid-svg-icons'
+    | 'phosphor/bold'
+    | 'phosphor/duotone'
+    | 'phosphor/fill'
+    | 'phosphor/light'
+    | 'phosphor/regular'
+    | 'phosphor/thin'
 
   /** Name of the icon according to the chosen library */
   selector: string